# services/api-gateway/generated/auth_client/openapi_client/models/independent_driver_register.py
# coding: utf-8

"""
    Auth Service

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json

from pydantic import BaseModel, ConfigDict, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
from typing import Optional, Set
from typing_extensions import Self

class IndependentDriverRegister(BaseModel):
    """
    IndependentDriverRegister
    """ # noqa: E501
    fname: StrictStr
    mname: Optional[StrictStr] = None
    lname: StrictStr
    email: StrictStr
    phone: StrictStr
    password: StrictStr
    license_number: Optional[StrictStr] = None
    __properties: ClassVar[List[str]] = ["fname", "mname", "lname", "email", "phone", "password", "license_number"]

    model_config = ConfigDict(
        populate_by_name=True,
        validate_assignment=True,
        protected_namespaces=(),
    )


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Optional[Self]:
        """Create an instance of IndependentDriverRegister from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        excluded_fields: Set[str] = set([
        ])

        _dict = self.model_dump(
            by_alias=True,
            exclude=excluded_fields,
            exclude_none=True,
        )
        # set to None if mname (nullable) is None
        # and model_fields_set contains the field
        if self.mname is None and "mname" in self.model_fields_set:
            _dict['mname'] = None

        # set to None if license_number (nullable) is None
        # and model_fields_set contains the field
        if self.license_number is None and "license_number" in self.model_fields_set:
            _dict['license_number'] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
        """Create an instance of IndependentDriverRegister from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "fname": obj.get("fname"),
            "mname": obj.get("mname"),
            "lname": obj.get("lname"),
            "email": obj.get("email"),
            "phone": obj.get("phone"),
            "password": obj.get("password"),
            "license_number": obj.get("license_number")
        })
        return _obj