# services/api-gateway/generated/auth_client/openapi_client/models/user_read.py
# coding: utf-8

"""
    Auth Service

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json

from datetime import datetime
from pydantic import BaseModel, ConfigDict, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
from openapi_client.models.role_read import RoleRead
from typing import Optional, Set
from typing_extensions import Self

class UserRead(BaseModel):
    """
    UserRead
    """ # noqa: E501
    id: StrictInt
    fname: StrictStr
    mname: Optional[StrictStr]
    lname: StrictStr
    email: StrictStr
    phone: StrictStr
    status: StrictStr
    role: RoleRead
    created_at: datetime
    updated_at: datetime
    __properties: ClassVar[List[str]] = ["id", "fname", "mname", "lname", "email", "phone", "status", "role", "created_at", "updated_at"]

    model_config = ConfigDict(
        populate_by_name=True,
        validate_assignment=True,
        protected_namespaces=(),
    )


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Optional[Self]:
        """Create an instance of UserRead from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        excluded_fields: Set[str] = set([
        ])

        _dict = self.model_dump(
            by_alias=True,
            exclude=excluded_fields,
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of role
        if self.role:
            _dict['role'] = self.role.to_dict()
        # set to None if mname (nullable) is None
        # and model_fields_set contains the field
        if self.mname is None and "mname" in self.model_fields_set:
            _dict['mname'] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
        """Create an instance of UserRead from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "id": obj.get("id"),
            "fname": obj.get("fname"),
            "mname": obj.get("mname"),
            "lname": obj.get("lname"),
            "email": obj.get("email"),
            "phone": obj.get("phone"),
            "status": obj.get("status"),
            "role": RoleRead.from_dict(obj["role"]) if obj.get("role") is not None else None,
            "created_at": obj.get("created_at"),
            "updated_at": obj.get("updated_at")
        })
        return _obj