# services/contracts/auth.openapi.json
{"openapi":"3.1.0","info":{"title":"Auth Service","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/info":{"get":{"summary":"Info","operationId":"info_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1":{"get":{"summary":"Root","operationId":"root_api_v1_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/register/passenger":{"post":{"tags":["Auth"],"summary":"Register Passenger","operationId":"register_passenger_api_v1_auth_register_passenger_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PassengerRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/register/vendor-admin":{"post":{"tags":["Auth"],"summary":"Register Vendor Admin","operationId":"register_vendor_admin_api_v1_auth_register_vendor_admin_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VendorAdminRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/register/independent-driver":{"post":{"tags":["Auth"],"summary":"Register Independent Driver","operationId":"register_independent_driver_api_v1_auth_register_independent_driver_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndependentDriverRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginSchema"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IndependentDriverRegister":{"properties":{"fname":{"type":"string","title":"Fname"},"mname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mname"},"lname":{"type":"string","title":"Lname"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"type":"string","title":"Phone"},"password":{"type":"string","title":"Password"},"license_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"License Number"}},"type":"object","required":["fname","lname","email","phone","password"],"title":"IndependentDriverRegister"},"LoginSchema":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginSchema"},"PassengerRegister":{"properties":{"fname":{"type":"string","title":"Fname"},"mname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mname"},"lname":{"type":"string","title":"Lname"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"type":"string","title":"Phone"},"password":{"type":"string","title":"Password"}},"type":"object","required":["fname","lname","email","phone","password"],"title":"PassengerRegister"},"RoleRead":{"properties":{"name":{"type":"string","title":"Name"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","id"],"title":"RoleRead"},"UserRead":{"properties":{"id":{"type":"integer","title":"Id"},"fname":{"type":"string","title":"Fname"},"mname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mname"},"lname":{"type":"string","title":"Lname"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"type":"string","title":"Phone"},"status":{"type":"string","title":"Status"},"role":{"$ref":"#/components/schemas/RoleRead"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","fname","mname","lname","email","phone","status","role","created_at","updated_at"],"title":"UserRead"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VendorAdminRegister":{"properties":{"fname":{"type":"string","title":"Fname"},"mname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mname"},"lname":{"type":"string","title":"Lname"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"type":"string","title":"Phone"},"password":{"type":"string","title":"Password"},"company_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Name"}},"type":"object","required":["fname","lname","email","phone","password"],"title":"VendorAdminRegister"}}}}